Broke down SupportRoomHelper.js into many smaller helper classes.
