+++ src/Helper/SupportRoomNameGenerator.js
@@ -1,0 +1,7 @@
+const crypto = require('crypto');
+
+const generateRoomName = (typeKey) => {
+    return `support-${typeKey}-${crypto.randomBytes(4).toString('hex')}`
+}
+
+module.exports = generateRoomName;
